+++ changelog.md
@@ -5,6 +5,7 @@
 - Fixed output files not having the correct path to the default stylesheet.
 - Fixed various issues with incorrect outputted file names.
 - Fixed `||` not being treated as valid table cell syntax.
+- Fixed front matter being generated even when config option `eleventy` was set to `false`.
 
 ## 1.3.1
 *2024-05-03*